requirements.txt
tqdm
transformers
sentence-transformers
gradio==3.41.2
av

pandas
huggingface-hub
Pillow
deepdanbooru
numpy
opencv-python

einops
imageio
imgaug
git+https://github.com/cocodataset/panopticapi.git
pytorch-lightning
albumentations==1.1.0
huggingface_hub
omegaconf
diffusers
accelerate
matplotlib
scipy
basicsr
numba
imgaug
moviepy
flask
h5py